Refactor code, improve clarity, and modify HTTP handling
This commit includes a series of changes as part of a larger refactoring effort. The `Code Metrics.xml` file was added to enable the DevExpress Code Analysis Metrics engine for all languages. In `FeeBayPlugin.cs`, code related to settings, logging, and tasks was commented out and a new `ScheduleTask` was added. The `DeleteProductAsync` method in `PollFeeBayTask.cs` was renamed to `DeleteProductOnDuckStampDealerAsync` for clarity, and unused code was removed. `HttpClient` usage was refactored in `Delete_InventoryItem.cs` to clear default headers before adding new ones. Import statements were modified in `FeeBayConnectionService.cs` to include `System.Collections.Generic`, unused code was removed, and a new method `GetFeeBayOrders` was added. Lastly, the `RequestAsync` method in `FeeBayRestHttpClient.cs` was removed.